backend/requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
drf-yasg
python-dotenv
djangorestframework-simplejwt
django-cors-headers